requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
bcrypt
psycopg2-binary
python-dotenv
